- standardization of getting csrf authentication token module
