+++ app/views/referer/show.html.erb
@@ -8,6 +8,7 @@
     user_name="<%= @declared_user %>"
     content="<%= @declare.declare %>"
     done_flg="<%= @declare.done %>"
+    v-bind:auth_token="auth_token"
   >
   </Refer>
